Extract node pool machine type into separate component
